feat: add fade-in / fade-out transition
- see https://stackoverflow.com/a/77715364
