+++ app/template.jsx
@@ -1,4 +1,4 @@
-import Transition from "@/components/ui/animation/transition";
+import Transition from "@/components/animation/transition";
 // see NextJS doc:
 // https://nextjs.org/docs/app/api-reference/file-conventions/template
 export default function Template({ children }) {
