+++ components/HOC/frozenRouter.jsx
@@ -1,0 +1,15 @@
+"use client";
+
+import { useContext, useRef } from "react";
+import { LayoutRouterContext } from "next/dist/shared/lib/app-router-context.shared-runtime";
+
+// Reference:
+// https://stackoverflow.com/a/77715364
+const FrozenRoute = ({ children }) => {
+  const context = useContext(LayoutRouterContext);
+  const frozen = useRef(context).current;
+
+  return <LayoutRouterContext.Provider value={frozen}>{children}</LayoutRouterContext.Provider>;
+};
+
+export default FrozenRoute;
