+++ components/animation/text-decode-cycle.jsx
