+++ components/animation/text-decode.jsx
